:bug: (Route: /api/v0.1/countries/population/filter) Fixed crashing.
Wrapped route code in a try/catch block. Errors now go to Express' error handler and result in an HTTP 500 instead of a
crash. Fixed an issue with sorting when population data doesn't exist for a given year. Data for countries without
population data for a queried year are now omitted from the results returned.