Bugfix/prevent hard crashes (#92)
* :bug: (Route: /api/v0.1/countries/population/filter) Fixed crashing.

Wrapped route code in a try/catch block. Errors now go to Express' error handler and result in an HTTP 500 instead of a
crash. Fixed an issue with sorting when population data doesn't exist for a given year. Data for countries without
population data for a queried year are now omitted from the results returned.

* :art: (Route: /api/v0.1/countries) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/cities) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/codes) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/codes) Cohesive route format

POST version of route. Wrapped the route function in a try / catch block that forwards errors to Express' built in error
handler via next()

* :art: (Route: /api/v0.1/countries/iso) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/currency) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/info) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next().
Also corrected response message from 'unicode' to 'unicodeFlag' if no GET parameters are supplied.

* :art: (Route: /api/v0.1/countries/capital) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/flag/images) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/flag/unicode) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/positions) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/positions/range) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next().
Added some TODO's to improve function.

* :art: (Route: /api/v0.1/countries/population) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next().
Added a TODO to improve function.

* :art: (Route: /api/v0.1/countries/population/cities) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :bug: (Route: /api/v0.1/countries/population/cities/filter) Fixed bug

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next().
Fixed a bug that would cause a crash if data for a specified city did not exist. Now if the results are empty a 404 is
returned.

* :art: (Route: /api/v0.1/countries/states) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/state/cities) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

* :art: (Route: /api/v0.1/countries/random) Cohesive route format

Wrapped the route function in a try / catch block that forwards errors to Express' built in error handler via next()

Closes #91